+++ src/index.ts
@@ -1,10 +1,10 @@
 import { DeepPartial } from "./utils/DeepPartial";
-import { CommandSetOptions } from "./models/ParseOptions";
+import { CommandSetOptions } from "./models/CommandSetOptions";
 
 export { makeCommand } from "./other/makeCommand";
 export { Command } from "./models/Command";
 export { CommandSet } from "./models/CommandSet";
-export { CommandSetOptions as ParseOptions } from "./models/ParseOptions";
+export { CommandSetOptions as ParseOptions } from "./models/CommandSetOptions";
 export { CommandResult } from "./models/CommandResult";
 export { Localization } from "./models/localization/Localization";
