+++ src/other/HelpUtils.ts
@@ -9,7 +9,7 @@ import { ArgumentRawHelp } from "../models/data/help/ArgumentRawHelp";
 import { FlagRawHelp } from "../models/data/help/FlagRawHelp";
 import { RestRawHelp } from "../models/data/help/RestRawHelp";
 import { CommandLocalization } from "../models/localization/CommandLocalization";
-import { CommandSetOptions } from "../models/ParseOptions";
+import { CommandSetOptions } from "../models/CommandSetOptions";
 import { isArray } from "../utils/array";
 import { template } from "../utils/template";
 import { reply } from "../utils/reply";
